Small fix to etcd cookbook
while working on the improvement of postgresql cookbook (especially for
easier HA setup) by following the article below,

https://github.com/compose/governor.git

I would like to use `etcd` in postgres cookbook, so I simply add the
following into the postgres cookbook (add.rb),

include_recipe "etcd::add"

however it needs some tweaks to make it finally work with another
cookbook (which is `postgres` in my case), which were incorporated in this commit.

Please feel free to comment if those tweaks are valid or not.