Keep up Redisio cookbook with upstream
The OneOps redisio cookbook was mostly inherited from

https://github.com/brianbianco/redisio

and the latest version of `brianbianco/redisio` (upstream) is 2.4.2,
while the OneOps redis cookbook version is only 1.7.0 (June 2013).

This is an effort to sync up with the upstream without losing the
compatibility with OneOps. For example, the updated code base will
still seamlessly support Redis `cluster` deployment with version 3.0.1

Note that the upstream cookbook supports `Redis Sentinal`, while the
OneOps redis cookbook did not. In this patch the sentinal related code
are checked in (to align with upstream) but the OneOps redis pack still does
not support Redis Sentinal deployment. In case Redis Sentinal will be
expected in future, supporting it should not be very complex (as the upstream
cookbook already provided a skeleton of Redis Sentinal deployment).

Overall, this patch does not eliminate or downgrade any existing
features and supported case/deployment, but just wants to keep up the
existing redisio cookbook with upstream. More supported deployment case
(e.g. redis sentinal) and improvements could be addressed later.

Deployment tests have been done.