Governor PG & Etcd: support VM replace & update
Etcd: support VM replace and Etcd update on primary and secondary
clouds, so that bad VM could be replaced. Note: for Etcd, if there are n
nodes in primary clouds, the maximum number of VM could be replaced at
one shot is n/2 (if n is an odd) or n/2-1 (if n is an even). This is
because if the number of replacement is larger than that, it will make
the Etcd cluster unavailable (loss of quorum) so that `delete.rb` (a
part of Etcd `replace.rb`) will fail.

In general, Etcd secondary cloud and replacement support is based on:

(1) Etcd has to depend on hostname (which is reasonable because IP
address could be changed upon VM replacement and using IP address in
Etcd config may not be highly encouraged)

(2) No LB is in the deployment plan. Because we propose to use the
method of querying platform-level fqdn (host `fqdn`) to figure out which
compute is from primary clouds. If using LB, the result of querying fqdn
will be the IP of LB, rather the IP of computes.

Governor PG: Based on the assumption of Etcd depending on hostname and
no LB is involved, the postgresql-governor pack file got updated here.

Tested on primary and secondary cloud setup with query examples.