Couchbase (CB): add diagonstic_cache to monitor CB
`diagonstic_cache` has been dropped in
https://github.com/oneops/circuit-oneops-1/pull/28

Primarily becuase (1) the issue mentioned in above PR. (2) it contains the walmart
Graphite url link.

Given that Graphite pack is already available on OneOps and this patch
will totally remove the Walmart Graphite reference in. Plus,
`diagonstic_cache` is a crucial part for monitoring and alerting Couchbase.
Personally I think it may be helpful to bring it back after addressing
all concerns, so that couchbase pack on OneOps will be close to
production-ready.

Tested with `single` and `redundant` environments.